requirements.txt
annotated-types
anyio
asgiref
asttokens
async-timeout
Bottleneck
cachetools
certifi
cffi
charset-normalizer
cryptography
decorator
Django
django-extensions
djangorestframework
docstring_parser
dotenv
drf-yasg
exceptiongroup
executing
google-api-core
google-auth
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-genai
google-resumable-media
googleapis-common-protos
greenlet
grpc-google-iam-v1
grpcio
grpcio-status
h11
httpcore
httpx
httpx-sse
idna
inflection
ipython
jedi
jsonpatch
jsonpointer
langchain
langchain-core
langchain-google-vertexai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
MarkupSafe
matplotlib-inline
numexpr
numpy
orjson
ormsgpack
packaging
pandas
parso
pexpect
prompt_toolkit
proto-plus
protobuf
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
Pygments
pymssql
pyOpenSSL
python-dateutil
python-dotenv
pytz
PyYAML
requests
requests-toolbelt
rsa
shapely
six
sniffio
SQLAlchemy
sqlparse
stack-data
tenacity
traitlets
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
validators
vertexai
wcwidth
websockets
Werkzeug
xxhash
zstandard
